feat(dashboard): enhance item variations logic and UI
- Refactor useItemVariationsPageLogics to include tab state management
- Add handleTabChange, handleOpenItemComparison, and handleBack functions
- Update ItemVariationsPage to utilize new logic and improve readability
- Introduce VariationTab type in types.ts for better type safety
